refactor: move CSVCenterImporter to infrastructure layer
