+++ oxytcmri/interface/importers.py
@@ -5,8 +5,6 @@ from pathlib import Path
 from oxytcmri.domain.entities.mri import Atlas, MRIExam
 from oxytcmri.domain.entities.subject import Subject
 from oxytcmri.domain.ports.repositories import AtlasRepository, SubjectRepository, MRIExamRepository, Repository
-from oxytcmri.domain.entities.center import Center
-from oxytcmri.domain.ports.repositories import CenterRepository
 from oxytcmri.interface.repositories.nifti_folders_mri_exam_repository import NiftiFoldersMRIExamRepository
 
 
@@ -37,49 +35,6 @@ class Importer(ABC):
         repositories: list[Repository]
             The list of repositories from which the importer can choose the needed ones.
         """
-
-
-class CSVCenterImporter(Importer):
-    """
-    Import centers from a CSV file to the repository.
-
-    Attributes:
-    -----------
-        filepath: str
-            The path to the CSV file.
-        center_repository: CenterRepository
-            The repository to save the centers.
-    """
-
-    def __init__(self, filepath: str, center_repository: CenterRepository = None):
-        self.filepath = Path(filepath)
-        # Ensure that the CSV file exists
-        if not self.filepath.exists():
-            raise FileNotFoundError(f"CSV file not found: '{self.filepath}'.")
-
-        self.center_repository = center_repository
-
-    def register_repository(self, repositories: list[Repository]):
-        for repository in repositories:
-            if isinstance(repository, CenterRepository):
-                self.center_repository = repository
-
-    def import_data(self) -> None:
-        """
-        Import centers from the CSV file to the repository.
-
-        Returns:
-        --------
-        None
-        """
-        if self.center_repository is None:
-            raise ValueError("Center repository is not set.")
-
-        with open(self.filepath, mode='r') as file:
-            reader = csv.DictReader(file)
-            centers = [Center(id=int(row['id']), name=row['name']) for row in reader]
-
-        self.center_repository.save_centers(centers)
 
 
 class CSVAtlasImporter(Importer):
