+++ oxytcmri/tests/unit/infrastructure/test_importers.py
@@ -1,6 +1,7 @@
 import pytest
 
-from oxytcmri.interface.importers import CSVCenterImporter, CSVAtlasImporter, NiftiFoldersImporter
+from oxytcmri.interface.importers import CSVAtlasImporter, NiftiFoldersImporter
+from oxytcmri.infrastructure.importers.csv import CSVCenterImporter
 from oxytcmri.tests.fixtures import path_to_test_data_folder
 from oxytcmri.tests.unit.domain.mocks import MockEmptyCenterRepository, MockAtlasRepository, \
     MockInMemoryEmptyMRIRepository, \
